174.1. Making use of useRef(Section 12: Understanding Hooks in React)
U.S. rep is probably the easiest to hope to use out of all the ones we're going to look at.

So let's just get to it back inside of our editor at the very top.

I will import use RF.

And then right after we declare a piece of state, we'll say RF is use RF.

We're then going to take that RF object and assign it to one of the elements inside or one we we're

returning from this component.

So in our case, we want to get a reference to the most apparent element that we're returning from our

component.

That is the div with the class name of UI form.

With Signe Arraf, we put in RF equals RF.

And that's pretty much it.

Now, after our component is rendered for the very first time, we can get a reference to that div by

making use of RF dot current.

It is specifically the current property on the ref that is gonna give us a reference to that div.

So as a quick example of that, let's put in a console log right about that return.

Of RF dot current.

I'm gonna say this, look back over.

You'll notice that the very first time our component is rendered, we get a council log of undefined.

That is because we are console logging, the value of wrapped up current before we return, any ASX

deal before the ref has been bound to this div.

And before that, Dave has even been created inside of our Dom.

However, after that, any time that we re render our component refought current will refer to our top

level div.

So I can clear my console and then I will cause a re render by changing the state of my dropdown component.

We could do that very easily by just opening the dropdown.

I left in the other consular log of the actual element that was clicked on so I can go remove that or

come into out very quickly.

The backup you're at the document body event listener.

We'll just delete that really quick and we'll try that again.

It's going to open that thing.

And there we go.

We now have a direct reference to that top level element inside of our component.

Do we access that through refought current?

OK.

So we now know the element that most parent element inside of our component.

And we know the element was actually clicked on.

So to be able to make the distinction.

Between scenario number two and scenario number one and decide whether or not the body event listener

should do anything.