174.2. Making use of useRef(Section 12: Understanding Hooks in React)
U.S. rep is probably the easiest to hope to use out of all the ones we're going to look at.

So let's just get to it back inside of our editor at the very top.

I will import use RF.

And then right after we declare a piece of state, we'll say RF is use RF.

We're then going to take that RF object and assign it to one of the elements inside or one we we're

returning from this component.

So in our case, we want to get a reference to the most apparent element that we're returning from our

component.

That is the div with the class name of UI form.

With Signe Arraf, we put in RF equals RF.

And that's pretty much it.

Now, after our component is rendered for the very first time, we can get a reference to that div by

making use of RF dot current.

It is specifically the current property on the ref that is gonna give us a reference to that div.

So as a quick example of that, let's put in a console log right about that return.

Of RF dot current.

I'm gonna say this, look back over.

You'll notice that the very first time our component is rendered, we get a council log of undefined.

That is because we are console logging, the value of wrapped up current before we return, any ASX

deal before the ref has been bound to this div.

And before that, Dave has even been created inside of our Dom.

However, after that, any time that we re render our component refought current will refer to our top

level div.

So I can clear my console and then I will cause a re render by changing the state of my dropdown component.

We could do that very easily by just opening the dropdown.

I left in the other consular log of the actual element that was clicked on so I can go remove that or

come into out very quickly.

The backup you're at the document body event listener.

We'll just delete that really quick and we'll try that again.

It's going to open that thing.

And there we go.

We now have a direct reference to that top level element inside of our component.

Do we access that through refought current?

OK.

So we now know the element that most parent element inside of our component.

And we know the element was actually clicked on.

So to be able to make the distinction.

Between scenario number two and scenario number one and decide whether or not the body event listener

should do anything.
The last thing we have to do.

And removed that council log.

And then up inside the body event listener, we're going to add inside of here.

If.

Ref dot, current dot contains event dot target.

That's it.

That's how we do this.

If that is true, then we're going to return and not do anything else inside of here.

The report current dock contains is going to see whether or not the element was clicked on is inside

of our components.

If the element, though, was clicked on is inside of our component, then we're going to return early.

Otherwise, it wasn't inside the component.

So we should attempt to close the dropdown by doing a set open with false.

And that's it.

That's the whole solution.

Now, I just want to reiterate just very quickly here, that contains method belongs to all DOM elements

and allows us to check if one DOM element is contained inside of another.

Yes, let's save this and do one last test.

The back over here.

I'm not going to click on the dropdown to open it and I'll click on a option when I click on the option.

It is contained inside of our component, which means we're going to return early.

Click on option.

We return early.

We do not attempt to call set open with false.

And so all we do is let the open piece of state to the opposite of whatever it currently is.

Which will close it up.

And I should also, Bill, to open this thing up, click anywhere outside.

And in that case, the element that we clicked on was not contained inside of that ref.

So we did not return early and we ended up closing the dropdown.

Yes, that looks pretty good.

All right.

Well, that is pretty much it for this dropdown.

So let's take a quick pause right here and continue in just a moment.