175.1. Body Event Listener Cleanup(Section 12: Understanding Hooks in React)
Before we move on, there's one last thing I want to show you around this drop down component.

I can open up my app James File.

Here it is right here.

I'm looking at the app component.

We're going to add in a little bit of code to this thing very quickly to toggle the visibility of the

dropdown component entirely.

So when I say toggle, I'm not talking about opening or closing the dropdown.

I'm talking about whether or not we show the drop down at all.

We're going to add this code in very quickly.

And then as soon as we do, we're going to start to see a really mysterious error message start to appear.

OK, so I'm going to add in a new piece of state inside the app component called Show DropDown, and

I'll get a setter for that as well.

And I'll default that to be a value of true, which means that I do want to show the dropdown.

Then right about the drop down itself, I can add in a button.

But the text of Tuggle dropped down.

On the button, I'm going to add in on click event listener.

So whenever someone calls on this thing, we're going to flip the value of show, drop down to the opposite

of whatever it currently is.

So inside that arrow function right there, we will call set show dropdown and then pass in the opposite.

Of show down.

And then finally, I'm going to wrap the drop down right here with a set of curly braces.

And inside the curly braces, we're going to take a look at the show dropped down.

He said, state, if that is true, then I want to show the drop down.

So notice I got to put in a question mark right here for a ternary expression and a show dropped on

his false.

I'll put in a colon and then no.

Which means display nothing at all.

All we're doing here is toggling the visibility or the existence of the dropdown entirely.

OK.

Let's now save this and tested out.

I see the drop down by default, which is good.

I can still open and close it.

Open it.

Like off screen, everything still works as expected.

But take a look at this.

If I click on toggle dropdown, the dropdown goes away again as expected.

And now if I click anywhere else on the screen, we get a really nasty error message.

So what's going on here?

Well, let's take a careful look at the error message itself.

It says cannot read property contains of No.

And it's pointing out some code inside of our event.

Listener.

All right.

Let's look back over.

And take a look at that inside of our body event listener right here inside of our dropdown Dodgiest

file.

So the air message is saying that refought current is evaluating to know, as it says we are trying

to call it, property contains on.

No, the report current is null instead of a reference to a DOM element, as we would expect.

So what exactly is happening?

Well, whenever we remove a component from the DOM, all the refs that are attached to elements inside

that component get set to note where more specifically the refought current property gets set to null

because we no longer have an element to refer to.

Remember this ref right here?

We are assigning to the top level element inside of our component.

So if that element right there is removed and the dumb ref current gets set to null because there is

no longer an element for us to refer to.

However, we still have this event listener set up this event, listeners still going to run whenever

we click on something inside the dumb.

So whenever we click on something Refco Current evaluates to know.

We then tried to call it a contains property on that.

And, well, we end up with that error message saying cannot refer to contains on known.