175.2. Body Event Listener Cleanup(Section 12: Understanding Hooks in React)
How are we going to fix this?

Well, what we really need to do here is make sure that whenever we are about to remove the drop down

component from the dam, we should turn off this event listener.

So to do so, we're going to make use of that cleanup function from use effect.

Once again, remember that we can return a function from USIE effect.

And this will be called right before the next time that this function is called.

There's one other scenario when this clean up function is going to be invoked.

And that is whenever we are about to stop showing the entire drop down component on the screen.

So if we are about to remove the drop down component entirely, this cleanup function will also be invoked.

That makes this a great location to do any kind of cleanup for the entire component that we want to

do instead of here.

Well, we really want to do is make sure that we do not attempt to call this event listener anymore.

We really want to detach or remove this event listener from the body element entirely.

And that's exactly what we're going to do to make sure that we remove this event.

Listener.

We're going to first kind of massage the code we have inside of your.

A little bit.

We're gonna take the callback itself.

So that's the event right there all the way down to the bottom of set open.

We will cut that and assign it to a new variable right above.

Hold on.

Body click.

And I'll paste in that function.

We can then provide on body click as the callback to the body click event listener.

And then finally, inside of our cleanup function, we're going to make sure that whenever our components

are about to be removed from the DOM, we remove that callback entirely.

We'll do a document dump body remove event listener.

Now, let's say that we want to remove an event listener from the click event and a specific event listener

we want to remove is the on body click function.

So now, whenever our component is first rendered into the DOM, because we've got this array right

here as the second argument, we're going to run the use effect function or the function we provide

to it.

Exactly one time at that point, we're going to setup the event listener.

Then whenever our dropdown is about to be removed from the DOM, react is going to automatically call

our cleanup function.

And that's going to remove the event.

Listener watching for that click.

All right.

Let's save this look back over.

Now, I can still open the things tonight, I'm close at all.

Everything is still working as expected.

And if I toggled visibility the down, I can then click anywhere on the screen and I don't get any error

messages.

As soon as I show the drop down again.

Well, it comes back up and I rebind that event.

Listener So use effect runs once again, event listener gets set up again and I can still open the dropdown,

click outside of it and see the thing close.

All right.

Well, that's it.

Well, that was pretty much it for our drop down component.

So we'll take a pause right here and continue on our next little widget in just a moment.